masm(chore): moved code around to reduce number of lines of codegen.rs
