+++ air-script/tests/system/system.masm
@@ -95,7 +95,7 @@ proc.cache_z_exp
     mem_load.4294903307 neg dup.0 neq.0
     while.true
         movdn.2 dup.1 dup.1 ext2mul
-        # => [z_1^2, z_0^2, i, ...]
+        # => [x_1^2, x_0^2, i, ...]
         movup.2 add.1 dup.0 neq.0
         # => [b, i+1, z_1^2, z_0^2, ...]
     end # END while
