+++ codegen/masm/src/codegen.rs
@@ -11,12 +11,13 @@ use ir::{
 };
 use std::collections::btree_map::{BTreeMap, Entry};
 use std::mem::{replace, take};
-use std::ops::{Bound, RangeBounds};
 
-use processor::math::{Felt, FieldElement, StarkField};
+use processor::math::{Felt, StarkField};
 use winter_prover::math::fft;
 
 use crate::config::CodegenConfig;
+use crate::error::CodegenError;
+use crate::exemption_points::gen_get_exemptions_points;
 use crate::writer::Writer;
 
 pub struct CodeGenerator<'ast> {
@@ -76,126 +77,6 @@ fn periodic_group_to_memory_offset(group: u32) -> u32 {
     // have to multiply the group by 2, to account for the zero padding, and add 1, to account for
     // the data being at the low and not high part of the word.
     group * 2 + 1
-}
-
-/// Loads the `element` from a memory range starting at `base_addr`.
-///
-/// This function is used to load a qudratic element from memory, and discard the other value. Even
-/// values are store in higher half of the word, while odd values are stored in the lower half.
-fn load_quadratic_element(
-    writer: &mut Writer,
-    base_addr: u32,
-    element: u32,
-) -> Result<(), CodegenError> {
-    let target_word: u32 = element / 2;
-    let address = base_addr + target_word;
-
-    // Load data from memory
-    writer.padw();
-    writer.mem_loadw(address);
-
-    // Discard the other value
-    match element % 2 {
-        0 => {
-            writer.movdn(3);
-            writer.movdn(3);
-            writer.drop();
-            writer.drop();
-        }
-        1 => {
-            writer.drop();
-            writer.drop();
-        }
-        _ => unreachable!(),
-    }
-
-    Ok(())
-}
-
-/// Precomputes the exemption points for a given power.
-///
-/// The current version of the generated code has hardcoded 2 exemption points, this means the
-/// points can be precomputed during compilation time to save a few cycles during runtime.
-fn points_for_power(power: u32) -> (u64, u64) {
-    let g = Felt::get_root_of_unity(power);
-    let trace_len = 2u64.pow(power);
-    let one = g.exp(trace_len - 1).as_int();
-    let two = g.exp(trace_len - 2).as_int();
-    (one, two)
-}
-
-/// Generate code to push the exemptions point to the top of the stack.
-///
-/// This procedure handles two power using conditional drops, instead of control flow with if
-/// statements, since the former is slightly faster for the small number of instructions used. The
-/// emitted code assumes the trace_length is at the top of the stack, and afer executing it will
-/// leave leave the stack as follows:
-///
-/// Stack: [g^{trace_len-2}, g^{trace_len-1}, ...]
-fn exemption_points(writer: &mut Writer, small_power: u32) {
-    let (lone, ltwo) = points_for_power(small_power);
-    let (hone, htwo) = points_for_power(small_power + 1);
-
-    writer.push(2u64.pow(small_power));
-    writer.u32checked_and();
-    writer.neq(0);
-    writer.comment(format!(
-        "Test if trace length is a power of 2^{}",
-        small_power
-    ));
-
-    writer.push(hone);
-    writer.push(lone);
-    writer.dup(2);
-    writer.cdrop();
-
-    writer.push(htwo);
-    writer.push(ltwo);
-    writer.movup(3);
-    writer.cdrop();
-}
-
-/// Helper function to emit efficient code to bisect the trace length value.
-///
-/// The callbacks `yes` and `no` are used to emit the code for each branch.
-fn bisect_trace_len<L, R>(writer: &mut Writer, range: impl RangeBounds<u32>, yes: L, no: R)
-where
-    L: FnOnce(&mut Writer),
-    R: FnOnce(&mut Writer),
-{
-    let mask = match (range.end_bound(), range.start_bound()) {
-        (Bound::Included(&start), Bound::Included(&end)) => {
-            let high_mask = 2u64.pow(start + 1) - 1;
-            let low_mask = 2u64.pow(end) - 1;
-            high_mask ^ low_mask
-        }
-        _ => panic!("Only inclusive ranges are supported"),
-    };
-
-    writer.dup(0);
-    writer.push(mask);
-    writer.u32checked_and();
-    writer.neq(0);
-    writer.comment(format!(
-        "{:?}..{:?}",
-        range.start_bound(),
-        range.end_bound()
-    ));
-
-    writer.r#if();
-    yes(writer);
-    writer.r#else();
-    no(writer);
-    writer.r#end();
-}
-
-/// Assumes a quadratic element is at the top of the stack and square it `n` times.
-fn quadratic_element_square(writer: &mut Writer, n: u32) {
-    for _ in 0..n {
-        writer.dup(1);
-        writer.dup(1);
-        writer.ext2mul();
-    }
 }
 
 impl<'ast> CodeGenerator<'ast> {
@@ -323,25 +204,10 @@ impl<'ast> CodeGenerator<'ast> {
             self.writer.header("=> [count, z_1, z_0, ...]");
 
             self.writer.header("Exponentiate z");
+            self.writer.exponentiate();
 
             // The trace length and the period may have the same size, so it is necessary to perform
             // the test before entering the loop.
-            self.writer.dup(0);
-            self.writer.neq(0);
-
-            self.writer.r#while();
-            self.writer.movdn(2);
-            self.writer.dup(1);
-            self.writer.dup(1);
-            self.writer.ext2mul();
-            self.writer.header("=> [z_1^2, z_0^2, i, ...]");
-
-            self.writer.movup(2);
-            self.writer.add(1);
-            self.writer.dup(0);
-            self.writer.neq(0);
-            self.writer.header("=> [b, i+1, z_1^2, z_0^2, ...]");
-            self.writer.end();
 
             let idx: u32 = idx.try_into().expect("periodic column length is too large");
             let addr = self.config.z_exp_address + idx;
@@ -377,22 +243,7 @@ impl<'ast> CodeGenerator<'ast> {
 
         // The trace length and the period may have the same size, so it is necessary to perform
         // the test before entering the loop.
-        self.writer.dup(0);
-        self.writer.neq(0);
-
-        self.writer.r#while();
-        self.writer.movdn(2);
-        self.writer.dup(1);
-        self.writer.dup(1);
-        self.writer.ext2mul();
-        self.writer.header("=> [z_1^2, z_0^2, i, ...]");
-
-        self.writer.movup(2);
-        self.writer.add(1);
-        self.writer.dup(0);
-        self.writer.neq(0);
-        self.writer.header("=> [b, i+1, z_1^2, z_0^2, ...]");
-        self.writer.end();
+        self.writer.exponentiate();
 
         let idx: u32 = self
             .periods
@@ -435,8 +286,7 @@ impl<'ast> CodeGenerator<'ast> {
 
         // `z^trace_len` is saved after all the period column points
         let group: u32 = self.periods.len().try_into().expect("periods are u32");
-        load_quadratic_element(
-            &mut self.writer,
+        self.writer.load_quadratic_element(
             self.config.z_exp_address,
             periodic_group_to_memory_offset(group),
         )?;
@@ -526,132 +376,11 @@ impl<'ast> CodeGenerator<'ast> {
     /// - The value `g^{trace_len-2}` is used both as an exemption point the integrity constraints
     /// divisor, as well as the offset point for the boundary constraints divisor.
     fn gen_get_exemptions_points(&mut self) -> Result<(), CodegenError> {
-        // Notes:
-        // - Computing the exemption points on the fly would require 1 exponentiation to find the
-        // root-of-unity from the two-adacity, followed by another exponetiation to compute the
-        // two-to-last exemption point, and a multiplication to compute the last exemption point.
-        // Each exponentiation is 41 cycles, giving around 83 cycles to compute the values.
-        // - For the range from powers 3 to 32 there are 30 unique values, which requires 8 words
-        // of data. Storing the data to memory requires pushing the 4 elements of a word to the
-        // stack, the target address, the store, and cleaning the stack, resulting in 10 cycles per
-        // word for a total of 80 cycles and some additional cycles to load the right value from
-        // memory when needed.
-        // - The code below instead uses a binary search to find the right value. And push only the
-        // necessary data to memory, in 62/73 cycles.
-        // - The smallest trace length is 2^3,
-        // Ref: https://github.com/facebook/winterfell/blob/main/air/src/air/trace_info.rs#L34-L35
-        // - The trace length is guaranteed to be a power-of-two and to fit in a u32.
-        // Ref: https://github.com/0xPolygonMiden/miden-vm/blob/next/stdlib/asm/crypto/stark/random_coin.masm#L76
-
         self.writer.proc("get_exemptions_points");
         self.load_trace_len();
         self.writer.header("=> [trace_len, ...]");
-
-        let writer = &mut self.writer;
-        bisect_trace_len(
-            writer,
-            3..=16,
-            |writer: &mut Writer| {
-                bisect_trace_len(
-                    writer,
-                    3..=10,
-                    |writer: &mut Writer| {
-                        bisect_trace_len(
-                            writer,
-                            3..=6,
-                            |writer: &mut Writer| {
-                                bisect_trace_len(
-                                    writer,
-                                    3..=4,
-                                    |writer: &mut Writer| exemption_points(writer, 3),
-                                    |writer: &mut Writer| exemption_points(writer, 5),
-                                )
-                            },
-                            |writer: &mut Writer| {
-                                bisect_trace_len(
-                                    writer,
-                                    7..=8,
-                                    |writer: &mut Writer| exemption_points(writer, 7),
-                                    |writer: &mut Writer| exemption_points(writer, 9),
-                                )
-                            },
-                        )
-                    },
-                    |writer: &mut Writer| {
-                        bisect_trace_len(
-                            writer,
-                            11..=14,
-                            |writer: &mut Writer| {
-                                bisect_trace_len(
-                                    writer,
-                                    11..=12,
-                                    |writer: &mut Writer| exemption_points(writer, 11),
-                                    |writer: &mut Writer| exemption_points(writer, 13),
-                                )
-                            },
-                            |writer: &mut Writer| exemption_points(writer, 15),
-                        )
-                    },
-                );
-            },
-            |writer: &mut Writer| {
-                bisect_trace_len(
-                    writer,
-                    17..=24,
-                    |writer: &mut Writer| {
-                        bisect_trace_len(
-                            writer,
-                            17..=20,
-                            |writer: &mut Writer| {
-                                bisect_trace_len(
-                                    writer,
-                                    17..=18,
-                                    |writer: &mut Writer| exemption_points(writer, 17),
-                                    |writer: &mut Writer| exemption_points(writer, 19),
-                                )
-                            },
-                            |writer: &mut Writer| {
-                                bisect_trace_len(
-                                    writer,
-                                    21..=22,
-                                    |writer: &mut Writer| exemption_points(writer, 21),
-                                    |writer: &mut Writer| exemption_points(writer, 23),
-                                )
-                            },
-                        )
-                    },
-                    |writer: &mut Writer| {
-                        bisect_trace_len(
-                            writer,
-                            25..=28,
-                            |writer: &mut Writer| {
-                                bisect_trace_len(
-                                    writer,
-                                    25..=26,
-                                    |writer: &mut Writer| exemption_points(writer, 25),
-                                    |writer: &mut Writer| exemption_points(writer, 27),
-                                )
-                            },
-                            |writer: &mut Writer| {
-                                bisect_trace_len(
-                                    writer,
-                                    29..=30,
-                                    |writer: &mut Writer| exemption_points(writer, 29),
-                                    |writer: &mut Writer| {
-                                        writer.drop();
-                                        let (one, two) = points_for_power(31);
-                                        writer.push(one);
-                                        writer.push(two);
-                                    },
-                                )
-                            },
-                        )
-                    },
-                );
-            },
-        );
-
-        self.writer.end(); // end proc
+        gen_get_exemptions_points(&mut self.writer)?;
+        self.writer.end();
 
         Ok(())
     }
@@ -732,18 +461,6 @@ impl<'ast> CodeGenerator<'ast> {
     }
 }
 
-#[derive(Debug)]
-pub enum CodegenError {
-    DuplicatedConstant,
-    InvalidAccessType,
-    UnknownConstant,
-    InvalidRowOffset,
-    InvalidSize,
-    InvalidIndex,
-    InvalidBoundaryConstraint,
-    InvalidIntegrityConstraint,
-}
-
 impl<'ast> AirVisitor<'ast> for CodeGenerator<'ast> {
     type Value = ();
     type Error = CodegenError;
@@ -822,8 +539,7 @@ impl<'ast> AirVisitor<'ast> for CodeGenerator<'ast> {
         self.writer
             .header("Multiply by the composition coefficient");
 
-        load_quadratic_element(
-            &mut self.writer,
+        self.writer.load_quadratic_element(
             self.config.composition_coef_address,
             self.composition_coefficient_count,
         )?;
@@ -865,8 +581,7 @@ impl<'ast> AirVisitor<'ast> for CodeGenerator<'ast> {
 
         // Note: The corretness of the load below relies on the integrity constraint being iterated
         // first _and_ the boundary constraints being iterated in natural order.
-        load_quadratic_element(
-            &mut self.writer,
+        self.writer.load_quadratic_element(
             self.config.composition_coef_address,
             self.composition_coefficient_count,
         )?;
@@ -924,7 +639,7 @@ impl<'ast> AirVisitor<'ast> for CodeGenerator<'ast> {
                 while e != 0 {
                     self.writer
                         .header(format!("square {} times", e.trailing_zeros()));
-                    quadratic_element_square(&mut self.writer, e.trailing_zeros());
+                    self.writer.quadratic_element_square(e.trailing_zeros());
 
                     // account for the exponentiations done above
                     e = e >> e.trailing_zeros();
@@ -983,8 +698,7 @@ impl<'ast> AirVisitor<'ast> for CodeGenerator<'ast> {
             .expect("All periods are added in the constructor")
             .try_into()
             .expect("periods are u32");
-        load_quadratic_element(
-            &mut self.writer,
+        self.writer.load_quadratic_element(
             self.config.z_exp_address,
             periodic_group_to_memory_offset(group),
         )?;
@@ -1110,7 +824,8 @@ impl<'ast> AirVisitor<'ast> for CodeGenerator<'ast> {
                     self.config.ood_aux_frame_address
                 };
 
-                load_quadratic_element(&mut self.writer, base_address, target_element)?;
+                self.writer
+                    .load_quadratic_element(base_address, target_element)?;
             }
             Value::PeriodicColumn(_, length) => {
                 let group: u32 = self
@@ -1120,8 +835,7 @@ impl<'ast> AirVisitor<'ast> for CodeGenerator<'ast> {
                     .expect("All periods are added in the constructor")
                     .try_into()
                     .expect("periods are u32");
-                load_quadratic_element(
-                    &mut self.writer,
+                self.writer.load_quadratic_element(
                     self.config.periodic_values_address,
                     periodic_group_to_memory_offset(group),
                 )?;
@@ -1139,15 +853,15 @@ impl<'ast> AirVisitor<'ast> for CodeGenerator<'ast> {
                 let index: u32 = (start_offset + *index)
                     .try_into()
                     .or(Err(CodegenError::InvalidIndex))?;
-                load_quadratic_element(&mut self.writer, self.config.public_inputs_address, index)?;
+                self.writer
+                    .load_quadratic_element(self.config.public_inputs_address, index)?;
             }
             Value::RandomValue(element) => {
                 // Compute the target address for the random value. Each memory address contains
                 // two values.
                 //
                 // Layout defined at: https://github.com/0xPolygonMiden/miden-vm/blob/next/stdlib/asm/crypto/stark/random_coin.masm#L169-L172
-                load_quadratic_element(
-                    &mut self.writer,
+                self.writer.load_quadratic_element(
                     self.config.aux_rand_address,
                     (*element).try_into().or(Err(CodegenError::InvalidIndex))?,
                 )?;
