+++ codegen/masm/src/error.rs
@@ -1,0 +1,11 @@
+#[derive(Debug)]
+pub enum CodegenError {
+    DuplicatedConstant,
+    InvalidAccessType,
+    UnknownConstant,
+    InvalidRowOffset,
+    InvalidSize,
+    InvalidIndex,
+    InvalidBoundaryConstraint,
+    InvalidIntegrityConstraint,
+}
